// src/domains/Launches/types/launches.ts
export interface Launches {
  flight_number: number;
  mission_name: string;
  mission_id: string[];
  launch_year: string;
  launch_date_unix: number;
  launch_date_utc: string;
  launch_date_local: string;
  is_tentative: boolean;
  tentative_max_precision: TentativeMaxPrecision;
  tbd: boolean;
  launch_window: number | null;
  rocket: Rocket;
  ships: string[];
  telemetry: Telemetry;
  launch_site: LaunchSite;
  launch_success: boolean | null;
  links: Links;
  details: null | string;
  upcoming: boolean;
  static_fire_date_utc: null | string;
  static_fire_date_unix: number | null;
  timeline: null;
  crew: any[] | null;
  last_date_update?: string;
  last_ll_launch_date?: null | string;
  last_ll_update?: null | string;
  last_wiki_launch_date?: string;
  last_wiki_revision?: string;
  last_wiki_update?: string;
  launch_date_source?: LaunchDateSource;
}

export enum LaunchDateSource {
  Wiki = "wiki",
}

export interface LaunchSite {
  site_id: SiteID;
  site_name: SiteName;
  site_name_long: SiteNameLong;
}

export enum SiteID {
  CcafsSlc40 = "ccafs_slc_40",
  KscLc39A = "ksc_lc_39a",
  VafbSlc4E = "vafb_slc_4e",
}

export enum SiteName {
  CcafsSlc40 = "CCAFS SLC 40",
  KscLc39A = "KSC LC 39A",
  VafbSlc4E = "VAFB SLC 4E",
}

export enum SiteNameLong {
  CapeCanaveralAirForceStationSpaceLaunchComplex40 = "Cape Canaveral Air Force Station Space Launch Complex 40",
  KennedySpaceCenterHistoricLaunchComplex39A = "Kennedy Space Center Historic Launch Complex 39A",
  VandenbergAirForceBaseSpaceLaunchComplex4E = "Vandenberg Air Force Base Space Launch Complex 4E",
}

export interface Links {
  mission_patch: null | string;
  mission_patch_small: undefined | string;
  reddit_campaign: string;
  reddit_launch: string;
  reddit_recovery: null | string;
  reddit_media: null | string;
  presskit: null;
  article_link: null | string;
  wikipedia: null | string;
  video_link: string;
  youtube_id: string;
  flickr_images: string[];
}

export interface Rocket {
  rocket_id: RocketID;
  rocket_name: RocketName;
  rocket_type: RocketType;
  first_stage: FirstStage;
  second_stage: SecondStage;
  fairings: Fairings | null;
}

export interface Fairings {
  reused: boolean | null;
  recovery_attempt: boolean | null;
  recovered: boolean | null;
  ship: null | string;
}

export interface FirstStage {
  cores: Core[];
}

export interface Core {
  core_serial: string;
  flight: number;
  block: number;
  gridfins: boolean;
  legs: boolean;
  reused: boolean;
  land_success: boolean;
  landing_intent: boolean;
  landing_type: LandingType;
  landing_vehicle: string;
}

export enum LandingType {
  Asds = "ASDS",
  Rtls = "RTLS",
}

export enum RocketID {
  Falcon9 = "falcon9",
}

export enum RocketName {
  Falcon9 = "Falcon 9",
}

export enum RocketType {
  Ft = "FT",
}

export interface SecondStage {
  block: number;
  payloads: Payload[];
}
// ...
export interface Telemetry {
  flight_club: null;
}

export enum TentativeMaxPrecision {
  Hour = "hour",
}